# chore(query): pick improve parse json performance to `release/v1.2.636-rc8.5`