--- a/.cargo/audit.toml
+++ b/.cargo/audit.toml
@@ -85,4 +85,6 @@ ignore = [
     "RUSTSEC-2025-0021",
     # openssl: Use-After-Free in `Md::fetch` and `Cipher::fetch`
     "RUSTSEC-2025-0022",
+    # Host panic with `fd_renumber` WASIp1 function
+    "RUSTSEC-2025-0046",
 ]