--- a/Cargo.toml
+++ b/Cargo.toml
@@ -245,7 +245,7 @@ geozero = { version = "0.13.0", features = ["default", "with-wkb", "with-geos", 
 hashbrown = { version = "0.14.3", default-features = false }
 http = "1"
 itertools = "0.10.5"
-jsonb = "0.5.2"
+jsonb = "0.5.3"
 jwt-simple = { version = "0.12.12", default-features = false, features = ["pure-rust"] }
 match-template = "0.0.1"
 mysql_async = { version = "0.34", default-features = false, features = ["native-tls-tls"] }
@@ -414,7 +414,7 @@ async-recursion = { git = "https://github.com/zhang2014/async-recursion.git", re
 backtrace = { git = "https://github.com/rust-lang/backtrace-rs.git", rev = "72265be" }
 color-eyre = { git = "https://github.com/eyre-rs/eyre.git", rev = "e5d92c3" }
 deltalake = { git = "https://github.com/delta-io/delta-rs", rev = "57795da" }
-jsonb = { git = "https://github.com/databendlabs/jsonb.git", rev = "edd714b" }
+jsonb = { git = "https://github.com/databendlabs/jsonb.git", rev = "6df281f" }
 openai_api_rust = { git = "https://github.com/datafuse-extras/openai-api", rev = "819a0ed" }
 # patched opendal which categories the XML dersierialization errors as recoverable
 opendal = { git = "https://github.com/datafuse-extras/opendal-for-release-v1.2.636", tag = "v0.49.0-xml-deser-retry" }