--- a/src/query/functions/src/scalars/binary.rs
+++ b/src/query/functions/src/scalars/binary.rs
@@ -31,6 +31,7 @@ use databend_common_expression::types::NumberType;
 use databend_common_expression::types::StringType;
 use databend_common_expression::types::UInt8Type;
 use databend_common_expression::types::ValueType;
+use databend_common_expression::types::VariantType;
 use databend_common_expression::Column;
 use databend_common_expression::EvalContext;
 use databend_common_expression::Function;
@@ -74,6 +75,15 @@ pub fn register(registry: &mut FunctionRegistry) {
         "try_to_string",
         |_, _| FunctionDomain::Full,
         error_to_null(eval_binary_to_string),
+    );
+
+    registry.register_passthrough_nullable_1_arg::<VariantType, BinaryType, _, _>(
+        "to_jsonb_binary",
+        |_, _| FunctionDomain::Full,
+        |val, _| match val {
+            ValueRef::Scalar(val) => Value::Scalar(val.to_vec()),
+            ValueRef::Column(col) => Value::Column(col),
+        },
     );
 
     registry.register_passthrough_nullable_1_arg::<StringType, BinaryType, _, _>(