--- a/src/query/functions/tests/it/scalars/testdata/binary.txt
+++ b/src/query/functions/tests/it/scalars/testdata/binary.txt
@@ -117,6 +117,42 @@ evaluation (internal):
 +--------+------------------------------------------------------------------------------------------------------------+
 
 
+ast            : to_jsonb_binary(parse_json('{"k1":"val","k2":100}'))
+raw expr       : to_jsonb_binary(parse_json('{"k1":"val","k2":100}'))
+checked expr   : to_jsonb_binary<Variant>(parse_json<String>("{\"k1\":\"val\",\"k2\":100}"))
+optimized expr : 40000002100000021000000210000003200000026B316B3276616C5064
+output type    : Binary
+output domain  : Undefined
+output         : 40000002100000021000000210000003200000026B316B3276616C5064
+
+
+ast            : to_jsonb_binary(parse_json('10'))
+raw expr       : to_jsonb_binary(parse_json('10'))
+checked expr   : to_jsonb_binary<Variant>(parse_json<String>("10"))
+optimized expr : 2000000020000002500A
+output type    : Binary
+output domain  : Undefined
+output         : 2000000020000002500A
+
+
+ast            : to_jsonb_binary(parse_json('123456'))
+raw expr       : to_jsonb_binary(parse_json('123456'))
+checked expr   : to_jsonb_binary<Variant>(parse_json<String>("123456"))
+optimized expr : 2000000020000005500001E240
+output type    : Binary
+output domain  : Undefined
+output         : 2000000020000005500001E240
+
+
+ast            : to_jsonb_binary(parse_json('"abcd"'))
+raw expr       : to_jsonb_binary(parse_json('"abcd"'))
+checked expr   : to_jsonb_binary<Variant>(parse_json<String>("\"abcd\""))
+optimized expr : 200000001000000461626364
+output type    : Binary
+output domain  : Undefined
+output         : 200000001000000461626364
+
+
 ast            : from_base64('QWJj')::String
 raw expr       : CAST(from_base64('QWJj') AS String)
 checked expr   : to_string<Binary>(from_base64<String>("QWJj"))