--- a/src/query/functions/tests/it/scalars/variant.rs
+++ b/src/query/functions/tests/it/scalars/variant.rs
@@ -77,6 +77,7 @@ fn test_parse_json(file: &mut impl Write) {
     run_ast(file, "parse_json('1234')", &[]);
     run_ast(file, "parse_json('[1,2,3,4]')", &[]);
     run_ast(file, "parse_json('{\"a\":\"b\",\"c\":\"d\"}')", &[]);
+    run_ast(file, "parse_json('{\"k\":\"v\",\"k\":\"v2\"}')", &[]);
 
     run_ast(file, "parse_json(s)", &[(
         "s",
@@ -112,6 +113,7 @@ fn test_try_parse_json(file: &mut impl Write) {
     run_ast(file, "try_parse_json('1234')", &[]);
     run_ast(file, "try_parse_json('[1,2,3,4]')", &[]);
     run_ast(file, "try_parse_json('{\"a\":\"b\",\"c\":\"d\"}')", &[]);
+    run_ast(file, "try_parse_json('{\"k\":\"v\",\"k\":\"v2\"}')", &[]);
 
     run_ast(file, "try_parse_json(s)", &[(
         "s",