@dantengsky 2025-08-04
NOTE: 

- this PR has been re-targeted to branch https://github.com/databendlabs/databend/compare/release/v1.2.636-rc8.6
- will be included in tag https://github.com/databendlabs/databend/tree/v1.2.636-rc8.6